# src/main/java/com/bank/yankiservice/dto/login/LoginRequest.java
package com.bank.yankiservice.dto.login;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@AllArgsConstructor
@NoArgsConstructor
@Data
public class LoginRequest {
    private String phoneNumber;
    private String documentNumber;
}